refactoring (each feature its folder, subdivided in model view services)
