+++ Code/educopter_frontend/analysis_options.yaml
@@ -24,6 +24,7 @@ linter:
   # producing the lint.
   rules:
     prefer_const_constructors : false
+    avoid_print: false
     # avoid_print: false  # Uncomment to disable the `avoid_print` rule
     # prefer_single_quotes: true  # Uncomment to enable the `prefer_single_quotes` rule
