+++ Code/educopter_frontend/lib/create_mission/model/dummy_data.dart
@@ -1,0 +1,9 @@
+  import 'package:educopter_frontend/select_activity/model/worldmap.dart';
+
+final List<Worldmap> availableWorldmaps = [
+    Worldmap(mapId: 1, mapName: 'Nederland', mapLocation: 'Netherlands.svg'),
+    Worldmap(mapId: 2, mapName: 'Duitsland', mapLocation: 'Duitsland.svg'),
+    Worldmap(mapId: 3, mapName: 'Europa', mapLocation: 'Europa.svg'),
+    Worldmap(mapId: 4, mapName: 'Zweden', mapLocation: 'Zweden.svg'),
+    Worldmap(mapId: 5, mapName: 'Verenigde Staten', mapLocation: 'VS.svg'),
+  ];
