+++ Code/educopter_frontend/lib/game/view/game.dart
