+++ Code/educopter_frontend/lib/general/general_widgets/maxcontentwidth.dart
