+++ Code/educopter_frontend/lib/select_activity/view/teacher_activity_options.dart
@@ -1,0 +1,37 @@
+import 'package:flutter/material.dart';
+
+class TeacherActivityOptions extends StatelessWidget {
+  const TeacherActivityOptions({
+    Key? key,
+    required this.userData,
+    required BuildContext context 
+  }) : super(key: key);
+
+  final Map userData;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+        color: Colors.green,
+        alignment: FractionalOffset.center,
+        child: (userData['rol'] != 'leerling')
+            ? Row(
+                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                children: [
+                  ElevatedButton(
+                    onPressed: (() {
+                      Navigator.of(context)
+                          .pushReplacementNamed('/createmission');
+                    }),
+                    child: Text('Create mission'),
+                  ),
+                  ElevatedButton(
+                    onPressed: (() {}),
+                    child: Text('Review students'),
+                  ),
+                ],
+              )
+            : null);
+  }
+}
+
